# packages/backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/purchases": {
      "post": {
        "operationId": "PurchasesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePurchaseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Purchases"
        ]
      },
      "get": {
        "operationId": "PurchasesController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Field to sort purchases by",
            "schema": {
              "default": "purchaseDate",
              "example": "purchaseDate",
              "type": "string",
              "enum": [
                "purchaseDate",
                "totalAmount",
                "createdAt",
                "patient.name",
                "consultant.name"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order",
            "schema": {
              "default": "asc",
              "example": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc",
                "ASC",
                "DESC"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term for purchases (e.g., patient name, product name - if implemented in service)",
            "schema": {
              "example": "Vitamin D",
              "type": "string"
            }
          },
          {
            "name": "patientId",
            "required": false,
            "in": "query",
            "description": "Filter purchases by patient ID",
            "schema": {
              "example": 12,
              "type": "number"
            }
          },
          {
            "name": "consultantId",
            "required": false,
            "in": "query",
            "description": "Filter purchases by consultant ID",
            "schema": {
              "example": 3,
              "type": "number"
            }
          },
          {
            "name": "purchaseDateFrom",
            "required": false,
            "in": "query",
            "description": "Filter purchases made on or after this date (YYYY-MM-DD)",
            "schema": {
              "example": "2023-05-01",
              "type": "string"
            }
          },
          {
            "name": "purchaseDateTo",
            "required": false,
            "in": "query",
            "description": "Filter purchases made on or before this date (YYYY-MM-DD)",
            "schema": {
              "example": "2023-05-31",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Purchases"
        ]
      }
    },
    "/api/purchases/{id}": {
      "get": {
        "operationId": "PurchasesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Purchases"
        ]
      }
    },
...
    "/api/statistics/total-revenue": {
      "get": {
        "description": "Calculates and returns the total revenue from all purchases and completed appointments. Also provides a breakdown of revenue from purchases and appointments separately.",
        "operationId": "StatisticsController_getTotalRevenue",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfully retrieved total revenue statistics.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "totalRevenue": {
                      "type": "number",
                      "example": 15250.75
                    },
                    "purchaseRevenue": {
                      "type": "number",
                      "example": 10000.5
                    },
                    "appointmentRevenue": {
                      "type": "number",
                      "example": 5250.25
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden resource. User does not have the right role."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get total revenue statistics",
        "tags": [
          "Statistics"
        ]
      }
    }
  },
  "info": {
    "title": "Supplement Management API",
    "description": "API documentation for the Supplement Management System",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreatePurchaseItemDto": {
        "type": "object",
        "properties": {
          "inventoryItemId": {
            "type": "number",
            "description": "ID of the inventory item being purchased",
            "example": 101
          },
          "quantity": {
            "type": "number",
            "description": "Quantity of the item being purchased, must be at least 1",
            "example": 2,
            "minimum": 1
          }
        },
        "required": [
          "inventoryItemId",
          "quantity"
        ]
      },
      "CreatePurchaseDto": {
        "type": "object",
        "properties": {
          "patientId": {
            "type": "number",
            "description": "Patient's ID for this purchase",
            "example": 7
          },
          "consultantId": {
            "type": "number",
            "description": "Consultant's ID who made the sale",
            "example": 2
          },
          "purchaseDate": {
            "type": "string",
            "description": "Date of purchase in YYYY-MM-DD format. Defaults to current date if not provided.",
            "example": "2023-10-26"
          },
          "items": {
            "description": "Array of items included in the purchase. Must contain at least one item.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreatePurchaseItemDto"
            }
          },
          "notes": {
            "type": "string",
            "description": "Optional notes for the purchase",
            "example": "Patient requested discrete packaging."
          }
        },
        "required": [
          "patientId",
          "consultantId",
          "items"
        ]
      },
...
    }
  }
}